py/pyproject.toml
[project]
name = "heaters-py"
version = "0.1.0"
description = "Python tasks for the Heaters media pipeline"
requires-python = ">=3.9"
dependencies = [
  "yt-dlp==2025.6.9",
  "Pillow>=10.1.0,<11.0.0",
  "numpy>=1.18.0,<2.0.0",
  "onnxruntime==1.20.1",
  "onnx_clip==4.0.1",
  "opencv-python-headless==4.11.0.86",
]

[tool.uv]
package = false
